refactor: keep spawn context in control plane
